[1단계 - 지하철 정보 관리 기능] 글렌(전석진) 미션 제출합니다. (#26)
* docs: 기능 구현 목록 추가


* feat: Section API 설계, 도메인 추가


* refactor: 컨트롤러 Request 요청 객체 분리


* feat: 데이터베이스 스카마 추가


* test: 기존 통합 테스트 코드 삭제

- 코드 리팩터링 및 추가에 방해


* feat: SectionDao, SectionEntity 추가


* feat: Sections 도메인 역 추가 기능 구현


* docs: 기능 구현 목록 체크


* chore: HTTP Request,Response 로그 제거


* chore: application.yml 추가


* feat: 데이터베이스에 노선에 구간을 추가하는 기능 추가


* refactor: 구간 추가 로직 리팩터링

- 검증 로직 도메인 이동
- 메소드 파라미터 순서 변경


* refactor: Station에 Id 삭제


* feat: Sections 생성 시 구간 정렬 기능 추가


* feat: 노선의 구간 목록을 조회하는 기능 추가


* feat: Line 도메인에 Id 제거, Line 상세 조회 기능 추가


* refactor: final 키워드 추가


* refactor: Request, Reponse 필드명 명확하게 변경


* build: logback 라이브러리 제거


* fix: 빈 구간의 노선 조회 시 빈 컬렉션 반환하게 수정


* feat: SectionService saveSection 로직 리팩터링


* feat: GlobalExceptionHandler 추가


* test: SectionController 통합 테스트 추가


* feat: 노선에 구간 삭제 기능 추가


* docs: 기능 요구 사항 체크


* feat: 전체 노선 조회 기능 구현


* refactor: dto 패키지 세부 분류


* refactor: StationNotFoundException 메시지 고정


* build: Spring Validation 추가


* feat: validation 적용


* refactor: LineDetailResponse 필드명 JsonProperty 적용


* docs: 기능 구현 목록 추가, 체크


* feat: 트랜잭션 적용


* refactor: deleteSection Section 변수명 명확하게 변경

* refactor: 구간을 추가할 때 노선에 존재하는 역이 있을 때 발생하는 예외 메시지 변경

* refactor: 구간 저장 시 역 검사 로직 리팩터링

- 있는지 조회하고 역을 만드는 것이 아닌, 데이터베이스에 있는 역을 가져온 뒤, 검사하도록 변경

* fix: 역, 노선을 저장할 때 기존의 값이 있는지 검사하는 로직 추가

* test: 테스트 이름 명확하게 변경

* feat: 도메인에 검증 로직 추가, 원시값 포장

* refactor: 상행, 하행역 용어 upBound, downBound로 변경

* feat: Entity 필드 접근자 final 추가

* test: 테스트용 sql 파일명 명확하게 수정

* docs: 요구 사항에 도메인 용어 수정

- startStation -> upBoundStation
- endStation -> downBoundStation

* chore: 코드 서식 수정

* feat: Response 응답 객체 추가, 응답 값 일관되게 변경

* fix: Station update SQL 수정

* test: StationDao 테스트 코드 추가

* feat: 삭제 시, 조회 시 없거나 있는 데이터를 조회할 때 검증 로직 추가

* test: LineDao 테스트 코드 추가

* build: Spring REST Docs 추가

* feat: GlobalExceptionHandler 예외 핸들러 추가

* feat: LineController 슬라이스 테스트 추가, RestDocs 적용

* build: asciidocter 플러그인 버전 수정

- 1.5.8 -> 2.4.0

* refactor: asciidoc 폴더 수정, 불필요한 테스트 삭제

* feat: EndsWith 커스텀 Validator 추가

* refactor: StationController showStaions 핸들러 메서드 이름 변경

- showStations -> findAllStations

* test: StationController 슬라이스 테스트 추가, RestDocs 적용

* test: stub 메서드 수정

* test: helper 패키지 추가, 픽스쳐 파일 분리

* test: SectionController 슬라이스 테스트 추가, RestDocs 적용

* feat: Section distance 필드, 원시값 포장

* chore: 오타 수정

* chore: .gitignore 파일 추가

* feat: SectionDao 인터페이스 교체, MemorySectionDao 구현체 추가

* test: SectionService 테스트 추가

---------

+++ src/main/java/subway/application/LineService.java
@@ -1,15 +1,23 @@
 package subway.application;
 
-import org.springframework.stereotype.Service;
-import subway.dao.LineDao;
-import subway.domain.Line;
-import subway.dto.LineRequest;
-import subway.dto.LineResponse;
+import static java.util.stream.Collectors.toList;
 
 import java.util.List;
-import java.util.stream.Collectors;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import subway.dao.LineDao;
+import subway.dao.entity.LineEntity;
+import subway.domain.Line;
+import subway.domain.LineColor;
+import subway.domain.LineName;
+import subway.dto.line.LineCreateRequest;
+import subway.dto.line.LineResponse;
+import subway.dto.line.LineUpdateRequest;
+import subway.exception.DuplicateLineException;
+import subway.exception.LineNotFoundException;
 
 @Service
+@Transactional
 public class LineService {
     private final LineDao lineDao;
 
@@ -17,37 +25,43 @@ public class LineService {
         this.lineDao = lineDao;
     }
 
-    public LineResponse saveLine(LineRequest request) {
-        Line persistLine = lineDao.insert(new Line(request.getName(), request.getColor()));
-        return LineResponse.of(persistLine);
+    public long saveLine(LineCreateRequest request) {
+        LineName lineName = new LineName(request.getLineName());
+        LineColor lineColor = new LineColor(request.getColor());
+        if (lineDao.existsByName(request.getLineName())) {
+            throw new DuplicateLineException();
+        }
+        return lineDao.insert(new Line(lineName, lineColor));
     }
 
+    @Transactional(readOnly = true)
     public List<LineResponse> findLineResponses() {
-        List<Line> persistLines = findLines();
-        return persistLines.stream()
-                .map(LineResponse::of)
-                .collect(Collectors.toList());
+        List<LineEntity> lines = lineDao.findAll();
+        return lines.stream()
+                .map(LineResponse::from)
+                .collect(toList());
     }
 
-    public List<Line> findLines() {
-        return lineDao.findAll();
-    }
-
+    @Transactional(readOnly = true)
     public LineResponse findLineResponseById(Long id) {
-        Line persistLine = findLineById(id);
-        return LineResponse.of(persistLine);
+        LineEntity lineEntity = lineDao.findById(id)
+                .orElseThrow(LineNotFoundException::new);
+        return new LineResponse(lineEntity.getId(), lineEntity.getName(), lineEntity.getColor());
     }
 
-    public Line findLineById(Long id) {
-        return lineDao.findById(id);
-    }
-
-    public void updateLine(Long id, LineRequest lineUpdateRequest) {
-        lineDao.update(new Line(id, lineUpdateRequest.getName(), lineUpdateRequest.getColor()));
+    public void updateLine(Long id, LineUpdateRequest request) {
+        LineName lineName = new LineName(request.getLineName());
+        LineColor lineColor = new LineColor(request.getColor());
+        if (lineDao.doesNotExistById(id)) {
+            throw new LineNotFoundException();
+        }
+        lineDao.update(id, new Line(lineName, lineColor));
     }
 
     public void deleteLineById(Long id) {
+        if (lineDao.doesNotExistById(id)) {
+            throw new LineNotFoundException();
+        }
         lineDao.deleteById(id);
     }
-
 }
